[#3347] Fix cppkakfa import on macos

+++ nes-core/src/Sources/SourceCreator.cpp
@@ -363,7 +363,7 @@ const DataSourcePtr createKafkaSource(SchemaPtr schema,
                                       OperatorId operatorId,
                                       OriginId originId,
                                       size_t numSourceLocalBuffers,
-                                      uint64_t batchSize,
+                                      size_t batchSize,
                                       const std::vector<Runtime::Execution::SuccessorExecutablePipeline>& successors) {
     return std::make_shared<KafkaSource>(schema,
                                          bufferManager,
